--- tests/data/latitude.json
@@ -1,241 +1,0 @@
-{
- "kind": "discovery#restDescription",
- "id": "latitude:v1",
- "name": "latitude",
- "version": "v1",
- "description": "Google Latitude API",
- "icons": {
-  "x16": "http://www.google.com/images/icons/product/search-16.gif",
-  "x32": "http://www.google.com/images/icons/product/search-32.gif"
- },
- "labels": [
-  "labs"
- ],
- "protocol": "rest",
- "basePath": "/latitude/v1/",
- "rootUrl": "https://www.googleapis.com/",
- "servicePath": "latitude/v1/",
- "auth": {
-  "oauth2": {
-   "scopes": {
-    "https://www.googleapis.com/auth/latitude": {
-     "description": "Manage your current location and location history"
-    }
-   }
-  }
- },
- "features": [
-  "dataWrapper"
- ],
- "schemas": {
-  "LatitudeCurrentlocationResourceJson": {
-   "$ref": "Location"
-  },
-  "Location": {
-   "id": "Location",
-   "type": "object",
-   "properties": {
-    "accuracy": {
-     "type": "any"
-    },
-    "activityId": {
-     "type": "any"
-    },
-    "altitude": {
-     "type": "any"
-    },
-    "altitudeAccuracy": {
-     "type": "any"
-    },
-    "heading": {
-     "type": "any"
-    },
-    "kind": {
-     "type": "string",
-     "default": "latitude#location"
-    },
-    "latitude": {
-     "type": "any"
-    },
-    "longitude": {
-     "type": "any"
-    },
-    "placeid": {
-     "type": "any"
-    },
-    "speed": {
-     "type": "any"
-    },
-    "timestampMs": {
-     "type": "any"
-    }
-   }
-  },
-  "LocationFeed": {
-   "id": "LocationFeed",
-   "type": "object",
-   "properties": {
-    "items": {
-     "type": "array",
-     "items": {
-      "$ref": "Location"
-     }
-    },
-    "kind": {
-     "type": "string",
-     "default": "latitude#locationFeed"
-    }
-   }
-  }
- },
- "resources": {
-  "currentLocation": {
-   "methods": {
-    "delete": {
-     "id": "latitude.currentLocation.delete",
-     "path": "currentLocation",
-     "httpMethod": "DELETE",
-     "description": "Deletes the authenticated user's current location.",
-     "scopes": [
-      "https://www.googleapis.com/auth/latitude"
-     ]
-    },
-    "get": {
-     "id": "latitude.currentLocation.get",
-     "path": "currentLocation",
-     "httpMethod": "GET",
-     "description": "Returns the authenticated user's current location.",
-     "parameters": {
-      "granularity": {
-       "type": "string",
-       "description": "Granularity of the requested location.",
-       "location": "query"
-      }
-     },
-     "response": {
-      "$ref": "LatitudeCurrentlocationResourceJson"
-     },
-     "scopes": [
-      "https://www.googleapis.com/auth/latitude"
-     ]
-    },
-    "insert": {
-     "id": "latitude.currentLocation.insert",
-     "path": "currentLocation",
-     "httpMethod": "POST",
-     "description": "Updates or creates the user's current location.",
-     "request": {
-      "$ref": "LatitudeCurrentlocationResourceJson"
-     },
-     "response": {
-      "$ref": "LatitudeCurrentlocationResourceJson"
-     },
-     "scopes": [
-      "https://www.googleapis.com/auth/latitude"
-     ]
-    }
-   }
-  },
-  "location": {
-   "methods": {
-    "delete": {
-     "id": "latitude.location.delete",
-     "path": "location/{locationId}",
-     "httpMethod": "DELETE",
-     "description": "Deletes a location from the user's location history.",
-     "parameters": {
-      "locationId": {
-       "type": "string",
-       "description": "Timestamp of the location to delete (ms since epoch).",
-       "required": true,
-       "location": "path"
-      }
-     },
-     "parameterOrder": [
-      "locationId"
-     ],
-     "scopes": [
-      "https://www.googleapis.com/auth/latitude"
-     ]
-    },
-    "get": {
-     "id": "latitude.location.get",
-     "path": "location/{locationId}",
-     "httpMethod": "GET",
-     "description": "Reads a location from the user's location history.",
-     "parameters": {
-      "granularity": {
-       "type": "string",
-       "description": "Granularity of the location to return.",
-       "location": "query"
-      },
-      "locationId": {
-       "type": "string",
-       "description": "Timestamp of the location to read (ms since epoch).",
-       "required": true,
-       "location": "path"
-      }
-     },
-     "parameterOrder": [
-      "locationId"
-     ],
-     "response": {
-      "$ref": "Location"
-     },
-     "scopes": [
-      "https://www.googleapis.com/auth/latitude"
-     ]
-    },
-    "insert": {
-     "id": "latitude.location.insert",
-     "path": "location",
-     "httpMethod": "POST",
-     "description": "Inserts or updates a location in the user's location history.",
-     "request": {
-      "$ref": "Location"
-     },
-     "response": {
-      "$ref": "Location"
-     },
-     "scopes": [
-      "https://www.googleapis.com/auth/latitude"
-     ]
-    },
-    "list": {
-     "id": "latitude.location.list",
-     "path": "location",
-     "httpMethod": "GET",
-     "description": "Lists the user's location history.",
-     "parameters": {
-      "granularity": {
-       "type": "string",
-       "description": "Granularity of the requested locations.",
-       "location": "query"
-      },
-      "max-results": {
-       "type": "string",
-       "description": "Maximum number of locations to return.",
-       "location": "query"
-      },
-      "max-time": {
-       "type": "string",
-       "description": "Maximum timestamp of locations to return (ms since epoch).",
-       "location": "query"
-      },
-      "min-time": {
-       "type": "string",
-       "description": "Minimum timestamp of locations to return (ms since epoch).",
-       "location": "query"
-      }
-     },
-     "response": {
-      "$ref": "LocationFeed"
-     },
-     "scopes": [
-      "https://www.googleapis.com/auth/latitude"
-     ]
-    }
-   }
-  }
- }
-}
