--- tests/data/malformed.json
@@ -1,1 +1,0 @@
-{
