+++ tests/test_file.py
@@ -31,7 +31,7 @@ import stat
 import tempfile
 import unittest
 
-from googleapiclient.http import HttpMockSequence
+from http_mock import HttpMockSequence
 from oauth2client import GOOGLE_TOKEN_URI
 from oauth2client import file
 from oauth2client import locked_file
