+++ tests/test_jwt.py
@@ -35,7 +35,7 @@ try:
 except ImportError:
     from cgi import parse_qs
 
-from googleapiclient.http import HttpMockSequence
+from http_mock import HttpMockSequence
 from oauth2client import crypt
 from oauth2client.anyjson import simplejson
 from oauth2client.client import Credentials
