+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27
+envlist = py26, py27
 
 [testenv]
 deps = keyring
@@ -12,7 +12,7 @@ deps = keyring
 setenv = PYTHONPATH=../google_appengine
 
 [testenv:py26]
-commands = nosetests --ignore-files=test_oauth2client_appengine\.py
+commands = nosetests --ignore-files=test_appengine\.py
 
 [testenv:py27]
-commands = nosetests
+commands = nosetests --ignore-files=test_appengine\.py
